Ranking: (1) move components from ranking following the directory structure from host app
