+++ src/pages/Ranking/components/ConfirmDialog/ConfirmDialog.tsx
