+++ src/pages/Ranking/components/Spinner/Spinner.tsx
