+++ src/pages/Ranking/components/icons/CloseIcon.tsx
