+++ src/pages/Ranking/pages/RankingList/components/RankingListItem.tsx
@@ -1,11 +1,11 @@
 import { useState } from "react"
 import { useTranslation } from "react-i18next"
 import { useNavigate } from "react-router-dom"
-import { Ranking } from "../../../domain/types/v1api"
-import { competitionResultsPath } from "../../../domain/competitionLink.ts"
-import { formatDate } from "../../../domain/formatDate.ts"
-import DropdownMenu, { DropdownMenuItem } from "../../../components/DropdownMenu/DropdownMenu.tsx"
-import SettingsIcon from "../../../components/icons/SettingsIcon.tsx"
+import { Ranking } from "../../../../../domain/types/v1api"
+import { competitionResultsPath } from "../../../shared/competitionLink.ts"
+import { formatDate } from "../shared/formatDate.ts"
+import DropdownMenu, { DropdownMenuItem } from "./DropdownMenu/DropdownMenu.tsx"
+import SettingsIcon from "./icons/SettingsIcon.tsx"
 import ProcessEventDialog from "./ProcessEventDialog.tsx"
 
 interface RankingListItemProps {
