+++ src/pages/Ranking/pages/RankingList/components/icons/SettingsIcon.tsx
