+++ src/pages/Ranking/pages/RankingList/shared/formatDate.ts
