+++ src/pages/Ranking/shared/competitionLink.ts
