+++ src/pages/Ranking/shared/scoreOptions.ts
