+++ src/pages/Ranking/shared/scoringAlgorithm.ts
