+++ src/pages/Ranking/shared/statusScores.ts
